Fix utf-8 handling (hopefully fixes #4)
Couldn't get it to handle it consistently, but this seems to work, we decode the
UTF-8 coming in, but it appears to we need to encode the UTF-8 as ASCII when
sending. No idea why right now, but this seems to work for the time being.

Also added some more error handling and logging while debugging.