Parameterize the command to use (#5)
I don't use neovim or neovim-remote so I'm not sure how it works, but perhaps
this will help.

Specify -L to set the launcher (defaults to gvim)
Specify -R to set the remote command (also defaults to gvim)

The -L command will get a --servername and other options normally passed to
gvim, though, so I don't know if neovim will work with that.

This introduces dependencies to cmdline, but it's in tcllib so should be fine.